Use the `ValkyrieIndexer.for` factory to resolve indexer classes
Introduces indexer mixins corresponding to `Hyrax::Resource` and
`Hyrax::Metadata::Core`. Use them to construct the beginnings of a work indexer
for Valkyrie works.

Add a class registry for indexer/resource class pairs. Register the Work indexer
as an example.

I played with using `Concurrent::Map` for this registry, but decided that if
folks are writing to the registry after app initialization, unexpected behavior
is just possible fallout from that.